Added vim-visual-star-search plugin
This allows you to select some text using Vim's visual mode and then hit *
and # to search for it elsewhere in the file.  For example, hit V, select
a strange sequence of characters like "$! $!", and hit star.  You'll find
all other runs of "$! $!" in the file.

If you hit <leader>* ("\*" unless you changed the mapleader), vim will
recursively vimgrep for the word under the cursor or the visual selection.